+++ evalml/model_understanding/__init__.py
@@ -2,6 +2,7 @@ from .graphs import (
     binary_objective_vs_threshold,
     calculate_permutation_importance,
     confusion_matrix,
+    get_linear_coefficients,
     get_prediction_vs_actual_data,
     get_prediction_vs_actual_over_time_data,
     graph_binary_objective_vs_threshold,
@@ -12,10 +13,12 @@ from .graphs import (
     graph_prediction_vs_actual,
     graph_prediction_vs_actual_over_time,
     graph_roc_curve,
+    graph_t_sne,
     normalize_confusion_matrix,
     partial_dependence,
     precision_recall_curve,
-    roc_curve
+    roc_curve,
+    t_sne
 )
 from .prediction_explanations import (
     explain_prediction,