+++ evalml/tests/automl_tests/test_automl_search_classification.py
@@ -1,3 +1,4 @@
+import pickle
 import time
 from unittest.mock import MagicMock, patch
 
@@ -19,6 +20,10 @@ from evalml.objectives import (
     get_objective
 )
 from evalml.pipelines import (
+    GeneratedPipelineBinary,
+    GeneratedPipelineMulticlass,
+    GeneratedPipelineTimeSeriesBinary,
+    GeneratedPipelineTimeSeriesMulticlass,
     ModeBaselineBinaryPipeline,
     ModeBaselineMulticlassPipeline,
     MulticlassClassificationPipeline,
@@ -120,7 +125,7 @@ def test_binary_auto(X_y_binary):
     best_pipeline = automl.best_pipeline
     assert best_pipeline._is_fitted
     y_pred = best_pipeline.predict(X)
-    assert len(np.unique(y_pred)) == 2
+    assert len(np.unique(y_pred.to_series())) == 2
 
 
 def test_multi_auto(X_y_multi, multiclass_core_objectives):
@@ -131,7 +136,7 @@ def test_multi_auto(X_y_multi, multiclass_core_objectives):
     best_pipeline = automl.best_pipeline
     assert best_pipeline._is_fitted
     y_pred = best_pipeline.predict(X)
-    assert len(np.unique(y_pred)) == 3
+    assert len(np.unique(y_pred.to_series())) == 3
 
     objective_in_additional_objectives = next((obj for obj in multiclass_core_objectives if obj.name == objective.name), None)
     multiclass_core_objectives.remove(objective_in_additional_objectives)
@@ -714,3 +719,49 @@ def test_automl_supports_time_series_classification(mock_binary_fit, mock_multi_
 
         assert result['parameters']['Delayed Feature Transformer'] == configuration
         assert result['parameters']['pipeline'] == configuration
+
+
+@pytest.mark.parametrize("problem_type", [ProblemTypes.BINARY, ProblemTypes.MULTICLASS])
+@patch('evalml.pipelines.MulticlassClassificationPipeline.fit')
+@patch('evalml.pipelines.MulticlassClassificationPipeline.score')
+@patch('evalml.pipelines.BinaryClassificationPipeline.fit')
+@patch('evalml.pipelines.BinaryClassificationPipeline.score')
+def test_automl_pickle_generated_pipeline(mock_binary_score, mock_binary_fit, mock_multi_score, mock_multi_fit,
+                                          problem_type, X_y_binary, X_y_multi):
+    if problem_type == ProblemTypes.BINARY:
+        X, y = X_y_binary
+        pipeline = GeneratedPipelineBinary
+
+    elif problem_type == ProblemTypes.MULTICLASS:
+        X, y = X_y_multi
+        pipeline = GeneratedPipelineMulticlass
+
+    a = AutoMLSearch(X_train=X, y_train=y, problem_type=problem_type)
+    a.search()
+
+    for i, row in a.rankings.iterrows():
+        assert a.get_pipeline(row['id']).__class__ == pipeline
+        assert pickle.loads(pickle.dumps(a.get_pipeline(row['id'])))
+
+
+@pytest.mark.parametrize('problem_type', [ProblemTypes.TIME_SERIES_MULTICLASS, ProblemTypes.TIME_SERIES_BINARY])
+@patch('evalml.pipelines.TimeSeriesMulticlassClassificationPipeline.score')
+@patch('evalml.pipelines.TimeSeriesBinaryClassificationPipeline.score')
+@patch('evalml.pipelines.TimeSeriesMulticlassClassificationPipeline.fit')
+@patch('evalml.pipelines.TimeSeriesBinaryClassificationPipeline.fit')
+def test_automl_time_series_classification_pickle_generated_pipeline(mock_binary_fit, mock_multi_fit, mock_binary_score, mock_multiclass_score,
+                                                                     problem_type, X_y_binary, X_y_multi):
+    if problem_type == ProblemTypes.TIME_SERIES_BINARY:
+        X, y = X_y_binary
+        pipeline = GeneratedPipelineTimeSeriesBinary
+    else:
+        X, y = X_y_multi
+        pipeline = GeneratedPipelineTimeSeriesMulticlass
+
+    configuration = {"gap": 0, "max_delay": 0, 'delay_target': False, 'delay_features': True}
+    a = AutoMLSearch(X_train=X, y_train=y, problem_type=problem_type, problem_configuration=configuration)
+    a.search()
+
+    for i, row in a.rankings.iterrows():
+        assert a.get_pipeline(row['id']).__class__ == pipeline
+        assert pickle.loads(pickle.dumps(a.get_pipeline(row['id'])))