+++ evalml/tests/component_tests/test_lsa.py
@@ -3,6 +3,8 @@ import logging
 import numpy as np
 import pandas as pd
 import pytest
+import woodwork as ww
+from pandas.testing import assert_frame_equal
 
 from evalml.pipelines.components import LSA
 
@@ -19,7 +21,7 @@ def test_lsa_only_text(text_df):
     X_t = lsa.transform(X)
     assert set(X_t.columns) == expected_col_names
     assert len(X_t.columns) == 4
-    assert X_t.dtypes.all() == np.float64
+    assert set(X_t.logical_types.values()) == {ww.logical_types.Double}
 
 
 def test_lsa_with_nontext(text_df):
@@ -36,7 +38,7 @@ def test_lsa_with_nontext(text_df):
     X_t = lsa.transform(X)
     assert set(X_t.columns) == expected_col_names
     assert len(X_t.columns) == 5
-    assert X_t.dtypes.all() == np.float64
+    assert set(X_t.logical_types.values()) == {ww.logical_types.Double}
 
 
 def test_lsa_no_text():
@@ -62,7 +64,7 @@ def test_some_missing_col_names(text_df, caplog):
     X_t = lsa.transform(X)
     assert set(X_t.columns) == expected_col_names
     assert len(X_t.columns) == 4
-    assert X_t.dtypes.all() == np.float64
+    assert set(X_t.logical_types.values()) == {ww.logical_types.Double}
 
 
 def test_all_missing_col_names(text_df):
@@ -115,7 +117,7 @@ def test_index_col_names():
     X_t = lsa.transform(X)
     assert set(X_t.columns) == expected_col_names
     assert len(X_t.columns) == 4
-    assert X_t.dtypes.all() == np.float64
+    assert set(X_t.logical_types.values()) == {ww.logical_types.Double}
 
 
 def test_int_col_names():
@@ -136,7 +138,7 @@ def test_int_col_names():
     X_t = lsa.transform(X)
     assert set(X_t.columns) == expected_col_names
     assert len(X_t.columns) == 4
-    assert X_t.dtypes.all() == np.float64
+    assert set(X_t.logical_types.values()) == {ww.logical_types.Double}
 
 
 def test_lsa_output():
@@ -146,14 +148,13 @@ def test_lsa_output():
                  'Red, the blood of angry men - black, the dark of ages past']})
     lsa = LSA(text_columns=['lsa'])
     lsa.fit(X)
-
-    expected_features = [[0.832, 0.],
-                         [0., 1.],
-                         [0.832, 0.]]
+    expected_features = pd.DataFrame([[0.832, 0.],
+                                      [0., 1.],
+                                      [0.832, 0.]], columns=["LSA(lsa)[0]", "LSA(lsa)[1]"])
     X_t = lsa.transform(X)
     cols = [col for col in X_t.columns if 'LSA' in col]
     features = X_t[cols]
-    np.testing.assert_almost_equal(features, expected_features, decimal=3)
+    assert_frame_equal(expected_features, features.to_dataframe(), atol=1e-3)
 
 
 def test_lsa_with_custom_indices(text_df):
@@ -162,4 +163,4 @@ def test_lsa_with_custom_indices(text_df):
     lsa = LSA(text_columns=['col_1', 'col_2'])
     lsa.fit(X)
     X_t = lsa.transform(X)
-    assert not X_t.isnull().any().any()
+    assert not X_t.to_dataframe().isnull().any().any()