+++ evalml/tests/component_tests/test_target_encoder.py
@@ -3,12 +3,11 @@ from unittest.mock import patch
 import numpy as np
 import pandas as pd
 import pytest
+from pandas.testing import assert_frame_equal
 from pytest import importorskip
 
 from evalml.exceptions import ComponentNotYetFittedError
-from evalml.pipelines import RegressionPipeline
 from evalml.pipelines.components import TargetEncoder
-from evalml.preprocessing import split_data
 
 importorskip('category_encoders', reason='Skipping test because category_encoders not installed')
 
@@ -58,7 +57,7 @@ def test_null_values_in_dataframe():
                                'col_2': [0.526894, 0.526894, 0.526894, 0.6, 0.526894],
                                'col_3': [0.6, 0.6, 0.6, 0.6, 0.6, ]})
 
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
     encoder = TargetEncoder(handle_missing='return_nan')
     encoder.fit(X, y)
@@ -66,7 +65,7 @@ def test_null_values_in_dataframe():
     X_expected = pd.DataFrame({'col_1': [0.6, 0.6, 0.6, 0.6, np.nan],
                                'col_2': [0.526894, 0.526894, 0.526894, 0.6, 0.526894],
                                'col_3': [0.6, 0.6, 0.6, 0.6, 0.6, ]})
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
     encoder = TargetEncoder(handle_missing='error')
     with pytest.raises(ValueError, match='Columns to be encoded can not contain null'):
@@ -77,19 +76,20 @@ def test_cols():
     X = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
                       'col_2': ['2', '1', '1', '1', '1'],
                       'col_3': ["a", "a", "a", "a", "a"]})
+    X_expected = X.astype({'col_1': 'Int64', 'col_2': 'category', 'col_3': 'category'})
     y = pd.Series([0, 1, 1, 1, 0])
     encoder = TargetEncoder(cols=[])
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    pd.testing.assert_frame_equal(X, X_t)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
     encoder = TargetEncoder(cols=['col_2'])
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    X_expected = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
+    X_expected = pd.DataFrame({'col_1': pd.Series([1, 2, 1, 1, 2], dtype="Int64"),
                                'col_2': [0.60000, 0.742886, 0.742886, 0.742886, 0.742886],
-                               'col_3': ["a", "a", "a", "a", "a"]})
-    pd.testing.assert_frame_equal(X_t, X_expected, check_less_precise=True)
+                               'col_3': pd.Series(["a", "a", "a", "a", "a"], dtype="category")})
+    assert_frame_equal(X_expected, X_t.to_dataframe(), check_less_precise=True)
 
     encoder = TargetEncoder(cols=['col_2', 'col_3'])
     encoder.fit(X, y)
@@ -97,7 +97,7 @@ def test_cols():
     encoder2 = TargetEncoder()
     encoder2.fit(X, y)
     X_t2 = encoder2.transform(X)
-    pd.testing.assert_frame_equal(X_t, X_t2)
+    assert_frame_equal(X_t.to_dataframe(), X_t2.to_dataframe())
 
 
 def test_transform():
@@ -108,10 +108,10 @@ def test_transform():
     encoder = TargetEncoder()
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    X_expected = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
+    X_expected = pd.DataFrame({'col_1': pd.Series([1, 2, 1, 1, 2], dtype="Int64"),
                                'col_2': [0.6, 0.65872, 0.6, 0.65872, 0.65872],
                                'col_3': [0.504743, 0.504743, 0.504743, 0.6, 0.504743]})
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
 
 def test_smoothing():
@@ -123,26 +123,26 @@ def test_smoothing():
     encoder = TargetEncoder(smoothing=1)
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    X_expected = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
-                               'col_2': [2, 1, 1, 1, 1],
+    X_expected = pd.DataFrame({'col_1': pd.Series([1, 2, 1, 1, 2], dtype="Int64"),
+                               'col_2': pd.Series([2, 1, 1, 1, 1], dtype="Int64"),
                                'col_3': [0.742886, 0.742886, 0.742886, 0.742886, 0.6]})
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
     encoder = TargetEncoder(smoothing=10)
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    X_expected = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
-                               'col_2': [2, 1, 1, 1, 1],
+    X_expected = pd.DataFrame({'col_1': pd.Series([1, 2, 1, 1, 2], dtype="Int64"),
+                               'col_2': pd.Series([2, 1, 1, 1, 1], dtype="Int64"),
                                'col_3': [0.686166, 0.686166, 0.686166, 0.686166, 0.6]})
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
     encoder = TargetEncoder(smoothing=100)
     encoder.fit(X, y)
     X_t = encoder.transform(X)
-    X_expected = pd.DataFrame({'col_1': [1, 2, 1, 1, 2],
-                               'col_2': [2, 1, 1, 1, 1],
+    X_expected = pd.DataFrame({'col_1': pd.Series([1, 2, 1, 1, 2], dtype="Int64"),
+                               'col_2': pd.Series([2, 1, 1, 1, 1], dtype="Int64"),
                                'col_3': [0.676125, 0.676125, 0.676125, 0.676125, 0.6]})
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
 
 
 def test_get_feature_names():
@@ -157,20 +157,6 @@ def test_get_feature_names():
     np.testing.assert_array_equal(encoder.get_feature_names(), np.array(['col_1', 'col_2', 'col_3']))
 
 
-def test_custom_indices():
-    # custom regression pipeline
-    class MyTargetPipeline(RegressionPipeline):
-        component_graph = ['Imputer', 'Target Encoder', 'Linear Regressor']
-        custom_name = "Target Pipeline"
-
-    X = pd.DataFrame({"a": ["a", "b", "a", "a", "a", "c", "c", "c"], "b": [0, 1, 1, 1, 1, 1, 0, 1]})
-    y = pd.Series([0, 0, 0, 1, 0, 1, 0, 0], index=[7, 2, 1, 4, 5, 3, 6, 8])
-
-    x1, x2, y1, y2 = split_data(X, y, problem_type='binary')
-    tp = MyTargetPipeline({})
-    tp.fit(x2, y2)
-
-
 @patch('evalml.pipelines.components.transformers.transformer.Transformer.fit')
 def test_pandas_numpy(mock_fit, X_y_binary):
     X, y = X_y_binary
@@ -180,7 +166,7 @@ def test_pandas_numpy(mock_fit, X_y_binary):
     X_t = pd.DataFrame(X).reset_index(drop=True, inplace=False)
 
     encoder.fit(X, y)
-    pd.testing.assert_frame_equal(mock_fit.call_args[0][0], X_t)
+    assert_frame_equal(mock_fit.call_args[0][0], X_t)
 
     X_numpy = X.to_numpy()
     encoder.fit(X_numpy, y)