+++ evalml/tests/pipeline_tests/test_pipelines.py
@@ -6,6 +6,7 @@ import numpy as np
 import pandas as pd
 import pytest
 import woodwork as ww
+from pandas.testing import assert_frame_equal
 from skopt.space import Integer, Real
 
 from evalml.demos import load_breast_cancer, load_wine
@@ -19,6 +20,12 @@ from evalml.model_family import ModelFamily
 from evalml.objectives import FraudCost, Precision
 from evalml.pipelines import (
     BinaryClassificationPipeline,
+    GeneratedPipelineBinary,
+    GeneratedPipelineMulticlass,
+    GeneratedPipelineRegression,
+    GeneratedPipelineTimeSeriesBinary,
+    GeneratedPipelineTimeSeriesMulticlass,
+    GeneratedPipelineTimeSeriesRegression,
     MulticlassClassificationPipeline,
     PipelineBase,
     RegressionPipeline
@@ -49,6 +56,7 @@ from evalml.pipelines.utils import (
     _get_pipeline_base_class,
     generate_pipeline_code,
     get_estimators,
+    get_generated_pipeline_class,
     make_pipeline,
     make_pipeline_from_components
 )
@@ -849,7 +857,7 @@ def make_mock_multiclass_pipeline():
 @patch('evalml.pipelines.RegressionPipeline.predict')
 def test_score_regression_single(mock_predict, mock_fit, X_y_regression):
     X, y = X_y_regression
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     clf = make_mock_regression_pipeline()
     clf.fit(X, y)
     objective_names = ['r2']
@@ -862,7 +870,7 @@ def test_score_regression_single(mock_predict, mock_fit, X_y_regression):
 @patch('evalml.pipelines.RegressionPipeline.predict')
 def test_score_nonlinear_regression(mock_predict, mock_fit, nonlinear_regression_pipeline_class, X_y_regression):
     X, y = X_y_regression
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     clf = nonlinear_regression_pipeline_class({})
     clf.fit(X, y)
     objective_names = ['r2']
@@ -908,9 +916,9 @@ def test_score_multiclass_single(mock_predict, mock_fit, mock_encode, X_y_binary
 @patch('evalml.pipelines.MulticlassClassificationPipeline._encode_targets')
 @patch('evalml.pipelines.MulticlassClassificationPipeline.fit')
 @patch('evalml.pipelines.ComponentGraph.predict')
-def test_score_nonlinear_multiclass(mock_encode, mock_fit, mock_predict, nonlinear_multiclass_pipeline_class, X_y_multi):
+def test_score_nonlinear_multiclass(mock_predict, mock_fit, mock_encode, nonlinear_multiclass_pipeline_class, X_y_multi):
     X, y = X_y_multi
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     mock_encode.return_value = y
     clf = nonlinear_multiclass_pipeline_class({})
     clf.fit(X, y)
@@ -924,7 +932,7 @@ def test_score_nonlinear_multiclass(mock_encode, mock_fit, mock_predict, nonline
 @patch('evalml.pipelines.RegressionPipeline.predict')
 def test_score_regression_list(mock_predict, mock_fit, X_y_binary):
     X, y = X_y_binary
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     clf = make_mock_regression_pipeline()
     clf.fit(X, y)
     objective_names = ['r2', 'mse']
@@ -971,7 +979,7 @@ def test_score_multi_list(mock_predict, mock_fit, mock_encode, X_y_binary):
 def test_score_regression_objective_error(mock_predict, mock_fit, mock_objective_score, X_y_binary):
     mock_objective_score.side_effect = Exception('finna kabooom 💣')
     X, y = X_y_binary
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     clf = make_mock_regression_pipeline()
     clf.fit(X, y)
     objective_names = ['r2', 'mse']
@@ -1015,7 +1023,7 @@ def test_score_binary_objective_error(mock_predict, mock_fit, mock_objective_sco
 def test_score_nonlinear_binary_objective_error(mock_predict, mock_fit, mock_objective_score, mock_encode, nonlinear_binary_pipeline_class, X_y_binary):
     mock_objective_score.side_effect = Exception('finna kabooom 💣')
     X, y = X_y_binary
-    mock_predict.return_value = y
+    mock_predict.return_value = ww.DataColumn(y)
     mock_encode.return_value = y
     clf = nonlinear_binary_pipeline_class({})
     clf.fit(X, y)
@@ -1060,18 +1068,19 @@ def test_compute_estimator_features(mock_scaler, mock_ohe, mock_imputer, X_y_bin
     X, y = X_y_binary
     X = pd.DataFrame(X)
     X_expected = pd.DataFrame(index=X.index, columns=X.columns).fillna(0)
-    mock_imputer.return_value = X
-    mock_ohe.return_value = X
-    mock_scaler.return_value = X_expected
+    mock_imputer.return_value = ww.DataTable(X)
+    mock_ohe.return_value = ww.DataTable(X)
+    mock_scaler.return_value = ww.DataTable(X_expected)
+    X_expected = X_expected.astype("Int64")
 
     pipeline = logistic_regression_binary_pipeline_class({})
     pipeline.fit(X, y)
 
     X_t = pipeline.compute_estimator_features(X)
-    pd.testing.assert_frame_equal(X_t, X_expected)
+    assert_frame_equal(X_expected, X_t.to_dataframe())
     assert mock_imputer.call_count == 2
     assert mock_ohe.call_count == 2
-    assert mock_scaler.call_count == 1
+    assert mock_scaler.call_count == 2
 
 
 @patch('evalml.pipelines.components.Imputer.transform')
@@ -1080,17 +1089,17 @@ def test_compute_estimator_features(mock_scaler, mock_ohe, mock_imputer, X_y_bin
 @patch('evalml.pipelines.components.ElasticNetClassifier.predict')
 def test_compute_estimator_features_nonlinear(mock_en_predict, mock_rf_predict, mock_ohe, mock_imputer, X_y_binary, nonlinear_binary_pipeline_class):
     X, y = X_y_binary
-    mock_imputer.return_value = pd.DataFrame(X)
-    mock_ohe.return_value = pd.DataFrame(X)
-    mock_en_predict.return_value = pd.Series(np.ones(X.shape[0]))
-    mock_rf_predict.return_value = pd.Series(np.zeros(X.shape[0]))
-    X_expected = pd.DataFrame({'Random Forest': np.zeros(X.shape[0]), 'Elastic Net': np.ones(X.shape[0])})
+    mock_imputer.return_value = ww.DataTable(X)
+    mock_ohe.return_value = ww.DataTable(X)
+    mock_en_predict.return_value = ww.DataColumn(np.ones(X.shape[0]))
+    mock_rf_predict.return_value = ww.DataColumn(np.zeros(X.shape[0]))
+    X_expected_df = pd.DataFrame({'Random Forest': np.zeros(X.shape[0]), 'Elastic Net': np.ones(X.shape[0])})
 
     pipeline = nonlinear_binary_pipeline_class({})
     pipeline.fit(X, y)
-
     X_t = pipeline.compute_estimator_features(X)
-    pd.testing.assert_frame_equal(X_t, X_expected)
+
+    assert_frame_equal(X_expected_df, X_t.to_dataframe())
     assert mock_imputer.call_count == 2
     assert mock_ohe.call_count == 4
     assert mock_en_predict.call_count == 2
@@ -1315,7 +1324,7 @@ def test_hyperparameters_none(dummy_classifier_estimator_class):
 @patch('evalml.pipelines.components.Estimator.predict')
 def test_score_with_objective_that_requires_predict_proba(mock_predict, dummy_regression_pipeline_class, X_y_binary):
     X, y = X_y_binary
-    mock_predict.return_value = pd.Series([1] * 100)
+    mock_predict.return_value = ww.DataColumn(pd.Series([1] * 100))
     # Using pytest.raises to make sure we error if an error is not thrown.
     with pytest.raises(PipelineScoreError):
         clf = dummy_regression_pipeline_class(parameters={})
@@ -1459,7 +1468,7 @@ def test_clone_fitted(X_y_binary, logistic_regression_binary_pipeline_class):
         pipeline_clone.predict(X)
     pipeline_clone.fit(X, y)
     X_t_clone = pipeline_clone.predict_proba(X)
-    pd.testing.assert_frame_equal(X_t, X_t_clone)
+    assert_frame_equal(X_t.to_dataframe(), X_t_clone.to_dataframe())
 
 
 def test_nonlinear_clone_fitted(X_y_binary, nonlinear_binary_pipeline_class):
@@ -1476,7 +1485,7 @@ def test_nonlinear_clone_fitted(X_y_binary, nonlinear_binary_pipeline_class):
         pipeline_clone.predict(X)
     pipeline_clone.fit(X, y)
     X_t_clone = pipeline_clone.predict_proba(X)
-    pd.testing.assert_frame_equal(X_t, X_t_clone)
+    assert_frame_equal(X_t.to_dataframe(), X_t_clone.to_dataframe())
 
 
 def test_feature_importance_has_feature_names(X_y_binary, logistic_regression_binary_pipeline_class):
@@ -1638,7 +1647,7 @@ def test_targets_data_types_classification_pipelines(data_type, problem_type, ta
     for pipeline_class in pipeline_classes:
         pipeline = helper_functions.safe_init_pipeline_with_njobs_1(pipeline_class)
         pipeline.fit(X, y)
-        predictions = pipeline.predict(X, objective)
+        predictions = pipeline.predict(X, objective).to_series()
         assert set(predictions.unique()).issubset(unique_vals)
         predict_proba = pipeline.predict_proba(X)
         assert set(predict_proba.columns) == set(unique_vals)
@@ -1767,13 +1776,13 @@ def test_stacked_estimator_in_pipeline(problem_type, X_y_binary, X_y_multi, X_y_
     pipeline = StackedPipeline(parameters=parameters)
     pipeline.fit(X, y)
     comparison_pipeline.fit(X, y)
-    assert not np.isnan(pipeline.predict(X)).values.any()
+    assert not np.isnan(pipeline.predict(X).to_series()).values.any()
 
     pipeline_score = pipeline.score(X, y, [objective])[objective]
     comparison_pipeline_score = comparison_pipeline.score(X, y, [objective])[objective]
 
     if problem_type == ProblemTypes.BINARY or problem_type == ProblemTypes.MULTICLASS:
-        assert not np.isnan(pipeline.predict_proba(X)).values.any()
+        assert not np.isnan(pipeline.predict_proba(X).to_dataframe()).values.any()
         assert (pipeline_score <= comparison_pipeline_score)
     else:
         assert (pipeline_score >= comparison_pipeline_score)
@@ -2413,3 +2422,19 @@ def test_get_component(logistic_regression_binary_pipeline_class, nonlinear_bina
     assert pipeline.get_component('OneHot_RandomForest') == OneHotEncoder(top_n=4)
     assert pipeline.get_component('Random Forest') == RandomForestClassifier()
     assert pipeline.get_component('Logistic Regression') == LogisticRegressionClassifier()
+
+
+@pytest.mark.parametrize("problem_type,resulting_class",
+                         [(ProblemTypes.BINARY, GeneratedPipelineBinary),
+                          (ProblemTypes.MULTICLASS, GeneratedPipelineMulticlass),
+                          (ProblemTypes.REGRESSION, GeneratedPipelineRegression),
+                          (ProblemTypes.TIME_SERIES_BINARY, GeneratedPipelineTimeSeriesBinary),
+                          (ProblemTypes.TIME_SERIES_MULTICLASS, GeneratedPipelineTimeSeriesMulticlass),
+                          (ProblemTypes.TIME_SERIES_REGRESSION, GeneratedPipelineTimeSeriesRegression),
+                          ("invalid", None)])
+def test_get_generated_pipeline_class(problem_type, resulting_class):
+    if problem_type != "invalid":
+        assert get_generated_pipeline_class(problem_type) == resulting_class
+    else:
+        with pytest.raises(ValueError, match="not recognized"):
+            get_generated_pipeline_class(problem_type)